[cloud](topn-opt) support topn 2 phase read in cloud mode
