+++ be/src/olap/storage_engine.h
@@ -129,7 +129,14 @@ public:
         return _segment_meta_mem_tracker;
     }
 
+    void add_quering_rowset(RowsetSharedPtr rs);
+
+    RowsetSharedPtr get_quering_rowset(RowsetId rs_id);
+
 protected:
+    void _evict_querying_rowset();
+    void _evict_quring_rowset_thread_callback();
+
     int32_t _effective_cluster_id = -1;
     HeartbeatFlags* _heartbeat_flags = nullptr;
 
@@ -140,6 +147,7 @@ protected:
     std::unique_ptr<RowsetIdGenerator> _rowset_id_generator;
     std::unique_ptr<MemTableFlushExecutor> _memtable_flush_executor;
     std::unique_ptr<CalcDeleteBitmapExecutor> _calc_delete_bitmap_executor;
+    CountDownLatch _stop_background_threads_latch;
 
     // This mem tracker is only for tracking memory use by segment meta data such as footer or index page.
     // The memory consumed by querying is tracked in segment iterator.
@@ -147,6 +155,11 @@ protected:
     // is similar to `-2912341218700198079`. So, temporarily put it in experimental type tracker.
     // maybe have to use ColumnReader count as segment meta size.
     std::shared_ptr<MemTracker> _segment_meta_mem_tracker;
+
+    // Hold reference of quering rowsets
+    std::mutex _quering_rowsets_mutex;
+    std::unordered_map<RowsetId, RowsetSharedPtr, HashOfRowsetId> _querying_rowsets;
+    scoped_refptr<Thread> _evict_quering_rowset_thread;
 };
 
 class StorageEngine final : public BaseStorageEngine {
@@ -265,12 +278,6 @@ public:
     void add_async_publish_task(int64_t partition_id, int64_t tablet_id, int64_t publish_version,
                                 int64_t transaction_id, bool is_recover);
     int64_t get_pending_publish_min_version(int64_t tablet_id);
-
-    void add_quering_rowset(RowsetSharedPtr rs);
-
-    RowsetSharedPtr get_quering_rowset(RowsetId rs_id);
-
-    void evict_querying_rowset(RowsetId rs_id);
 
     bool add_broken_path(std::string path);
     bool remove_broken_path(std::string path);
@@ -406,14 +413,9 @@ private:
     PendingRowsetSet _pending_local_rowsets;
     PendingRowsetSet _pending_remote_rowsets;
 
-    // Hold reference of quering rowsets
-    std::mutex _quering_rowsets_mutex;
-    std::unordered_map<RowsetId, RowsetSharedPtr, HashOfRowsetId> _querying_rowsets;
-
     // Count the memory consumption of segment compaction tasks.
     std::shared_ptr<MemTracker> _segcompaction_mem_tracker;
 
-    CountDownLatch _stop_background_threads_latch;
     scoped_refptr<Thread> _unused_rowset_monitor_thread;
     // thread to monitor snapshot expiry
     scoped_refptr<Thread> _garbage_sweeper_thread;
