+++ regression-test/pipeline/cloud_p0/conf/session_variables.sql
@@ -3,4 +3,3 @@ set global insert_visible_timeout_ms=60000;
 set global enable_auto_analyze=false;
 set global enable_audit_plugin=true;
 set global enable_memtable_on_sink_node=false;
-set global enable_two_phase_read_opt = false;
